+++ sim-test.c
@@ -3088,6 +3088,47 @@ int test_local_effect_calculators(RND_U32 seed) {
 	assert(b.markers_in_block[1][1] == 3);
 	assert(b.markers_in_block[1][2] == 4);
     assert(b.markers_in_block[2][0] == 5);
+    
+    MarkerBlocks b2 = create_evenlength_blocks_each_chr(d, init.map, 2);
+    assert(b2.num_blocks == 6);
+    assert(b2.num_markers_in_block[0] == 1);
+    assert(b2.num_markers_in_block[1] == 1);
+    assert(b2.num_markers_in_block[2] == 2);
+    assert(b2.num_markers_in_block[3] == 1);
+    assert(b2.num_markers_in_block[4] == 1);
+    assert(b2.num_markers_in_block[5] == 0);
+    assert(b2.markers_in_block[0][0] == 0);
+    assert(b2.markers_in_block[1][0] == 1);
+    assert(b2.markers_in_block[2][0] == 2);
+    assert(b2.markers_in_block[2][1] == 3);
+    assert(b2.markers_in_block[3][0] == 4);
+    assert(b2.markers_in_block[4][0] == 5);
+    delete_markerblocks(&b2);
+    
+    b2 = create_evenlength_blocks_each_chr(d, init.map, 5);
+    assert(b2.num_blocks == 3*5);
+    assert(b2.num_markers_in_block[0] == 1);
+    assert(b2.num_markers_in_block[1] == 0);
+    assert(b2.num_markers_in_block[2] == 0);
+    assert(b2.num_markers_in_block[3] == 0);
+    assert(b2.num_markers_in_block[4] == 1);
+    assert(b2.num_markers_in_block[5] == 1);
+    assert(b2.num_markers_in_block[6] == 1);
+    assert(b2.num_markers_in_block[7] == 0);
+    assert(b2.num_markers_in_block[8] == 0);
+    assert(b2.num_markers_in_block[9] == 1);
+    assert(b2.num_markers_in_block[10] == 1);
+    assert(b2.num_markers_in_block[11] == 0);
+    assert(b2.num_markers_in_block[12] == 0);
+    assert(b2.num_markers_in_block[13] == 0);
+    assert(b2.num_markers_in_block[14] == 0);
+    assert(b2.markers_in_block[0][0] == 0);
+    assert(b2.markers_in_block[4][0] == 1);
+    assert(b2.markers_in_block[5][0] == 2);
+    assert(b2.markers_in_block[6][0] == 3);
+    assert(b2.markers_in_block[9][0] == 4);
+    assert(b2.markers_in_block[10][0] == 5);
+    delete_markerblocks(&b2);
 
 	DecimalMatrix dec = calculate_local_bvs(d, init.group, b, init.effSet);
 	assert(dec.dim2 == 3); // blocks