+++ NEWS.md
@@ -22,6 +22,7 @@ Latest News       {#news}
 - Fix a segfault that occured when calling `load_data_files` with `genotype_file = NULL`.
 - Fix a segfault that occured when trying to create iterators for a simulation object with no loaded genotypes.
 - Fix a bug where marker names had trailing spaces (and therefore, when loading more input files later, could not be matched) when automatically generating a genetic map from a genotype file. 
+- Fix a bug in `create_evenlength_blocks_each_chr` where too many markers would be allocated to the last block.
 
 # genomicSimulation v0.2.6